--- SpartaHack II/MentorFormCell.xib
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="11762" systemVersion="16B2659" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="11757"/>
-        <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="mentorFormCell" rowHeight="207" id="uRu-xB-yWI" customClass="MentorFormCell" customModule="SpartaHack_2016" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="530" height="207"/>
-            <autoresizingMask key="autoresizingMask"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="uRu-xB-yWI" id="0kK-jG-arj">
-                <rect key="frame" x="0.0" y="0.0" width="530" height="206.5"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" scrollEnabled="NO" text="Describe your problem. An example would be: &quot;Help! I can't figure out which awesome animation library to use for my web app!&quot;" textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="j1n-zr-gkY">
-                        <rect key="frame" x="16" y="84" width="498" height="69"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                        <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                        <textInputTraits key="textInputTraits" autocapitalizationType="sentences"/>
-                    </textView>
-                    <textField opaque="NO" clipsSubviews="YES" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" usesAttributedText="YES" borderStyle="roundedRect" placeholder="Category" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="yRq-Ui-mHQ">
-                        <rect key="frame" x="16" y="8" width="498" height="30"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="30" id="8ui-JE-eko"/>
-                        </constraints>
-                        <attributedString key="attributedText"/>
-                        <textInputTraits key="textInputTraits"/>
-                    </textField>
-                    <textField opaque="NO" clipsSubviews="YES" contentMode="scaleToFill" verticalHuggingPriority="251" verticalCompressionResistancePriority="751" contentHorizontalAlignment="left" contentVerticalAlignment="center" usesAttributedText="YES" borderStyle="roundedRect" placeholder="Location" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="s7c-eK-xwn">
-                        <rect key="frame" x="16" y="46" width="498" height="30"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="30" id="QK2-IN-nMf"/>
-                        </constraints>
-                        <attributedString key="attributedText"/>
-                        <textInputTraits key="textInputTraits"/>
-                    </textField>
-                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Hq5-AI-Uye">
-                        <rect key="frame" x="16" y="161" width="498" height="30"/>
-                        <constraints>
-                            <constraint firstAttribute="height" relation="greaterThanOrEqual" constant="30" id="B4L-1G-7YK"/>
-                            <constraint firstAttribute="height" constant="30" id="Eb8-do-B2Z"/>
-                        </constraints>
-                        <state key="normal" title="Submit Ticket"/>
-                    </button>
-                </subviews>
-                <constraints>
-                    <constraint firstAttribute="trailing" secondItem="s7c-eK-xwn" secondAttribute="trailing" constant="16" id="1xT-wh-8fR"/>
-                    <constraint firstItem="yRq-Ui-mHQ" firstAttribute="top" secondItem="0kK-jG-arj" secondAttribute="top" constant="8" id="3ai-Ie-oHp"/>
-                    <constraint firstItem="s7c-eK-xwn" firstAttribute="top" secondItem="yRq-Ui-mHQ" secondAttribute="bottom" constant="8" id="50D-qy-NyL"/>
-                    <constraint firstItem="j1n-zr-gkY" firstAttribute="top" secondItem="s7c-eK-xwn" secondAttribute="bottom" constant="8" id="FdY-6w-zy2"/>
-                    <constraint firstAttribute="trailing" secondItem="j1n-zr-gkY" secondAttribute="trailing" constant="16" id="NbW-hR-gEy"/>
-                    <constraint firstAttribute="trailing" secondItem="yRq-Ui-mHQ" secondAttribute="trailing" constant="16" id="RuI-Rw-Ujb"/>
-                    <constraint firstItem="Hq5-AI-Uye" firstAttribute="top" secondItem="j1n-zr-gkY" secondAttribute="bottom" constant="8" id="aOc-UP-9mn"/>
-                    <constraint firstAttribute="trailing" secondItem="Hq5-AI-Uye" secondAttribute="trailing" constant="16" id="aZG-SX-myA"/>
-                    <constraint firstItem="j1n-zr-gkY" firstAttribute="leading" secondItem="0kK-jG-arj" secondAttribute="leading" constant="16" id="brT-Xm-Ohw"/>
-                    <constraint firstItem="Hq5-AI-Uye" firstAttribute="leading" secondItem="0kK-jG-arj" secondAttribute="leading" constant="16" id="cX7-Fw-IdB"/>
-                    <constraint firstItem="yRq-Ui-mHQ" firstAttribute="leading" secondItem="0kK-jG-arj" secondAttribute="leading" constant="16" id="oy5-2a-DJc"/>
-                    <constraint firstAttribute="bottomMargin" secondItem="Hq5-AI-Uye" secondAttribute="bottom" constant="8" id="sNG-81-D8c"/>
-                    <constraint firstItem="s7c-eK-xwn" firstAttribute="leading" secondItem="0kK-jG-arj" secondAttribute="leading" constant="16" id="z3f-gA-fzh"/>
-                    <constraint firstItem="yRq-Ui-mHQ" firstAttribute="top" secondItem="0kK-jG-arj" secondAttribute="top" constant="8" id="zSP-WS-uwv"/>
-                </constraints>
-            </tableViewCellContentView>
-            <connections>
-                <outlet property="categoryTextField" destination="yRq-Ui-mHQ" id="CSC-20-ikX"/>
-                <outlet property="descriptionTextView" destination="j1n-zr-gkY" id="2RQ-re-Wux"/>
-                <outlet property="locationTextField" destination="s7c-eK-xwn" id="ds5-Ax-29S"/>
-                <outlet property="submitTicketButton" destination="Hq5-AI-Uye" id="ms9-xG-02S"/>
-            </connections>
-            <point key="canvasLocation" x="113" y="-33.5"/>
-        </tableViewCell>
-    </objects>
-</document>
