+++ SpartaHack II/MentorshipViewController.swift
@@ -6,9 +6,37 @@
 //  Copyright © 2016 Chris McGrath. All rights reserved.
 //
 
-import UIKit
+import Foundation
 
-class MentorshipViewController: SpartaTableViewController {
+class MentorshipViewController: UIViewController, UITextFieldDelegate, UITextViewDelegate {
+    @IBOutlet weak var categoryTextField: UITextField!
+    @IBOutlet weak var locationTextField: UITextField!
+    @IBOutlet weak var descriptionTextView: UITextView!
+    @IBOutlet weak var submitTicketButton: UIButton!
+    @IBOutlet weak var scrollView: UIScrollView!
+    @IBOutlet weak var contentView: UIView!
+    
+    var scrollViewContentInset: UIEdgeInsets = .zero
+    
+    var placeHolderText = "Describe your problem. An example would be: \"Help! I can't figure out which awesome animation library to use for my web app!\""
+    
+    func textViewShouldBeginEditing(_ textView: UITextView) -> Bool {
+        
+        self.descriptionTextView.textColor = Theme.primaryColor
+        
+        if(self.descriptionTextView.text == placeHolderText) {
+            self.descriptionTextView.text = ""
+        }
+        
+        return true
+    }
+    
+    func textViewDidEndEditing(_ textView: UITextView) {
+        if(textView.text == "") {
+            self.descriptionTextView.text = placeHolderText
+            self.descriptionTextView.textColor = Theme.tintColor
+        }
+    }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
@@ -16,56 +44,78 @@ class MentorshipViewController: SpartaTableViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.tableView.estimatedRowHeight = 350
-        self.tableView.alwaysBounceVertical = false // disable scrolling
-        let mentorFormNib = UINib(nibName: "MentorFormCell", bundle: Bundle(for: type(of: self)))
-        self.tableView.register(mentorFormNib, forCellReuseIdentifier: "mentorFormCell")
+        
+        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name:NSNotification.Name.UIKeyboardWillShow, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHide), name:NSNotification.Name.UIKeyboardWillHide, object: nil)
+
+        
         self.hideKeyboard()
     }
 
-    func tableView(tableView: UITableView, heightForRowAtIndexPath indexPath: NSIndexPath) -> CGFloat {
-        return UITableViewAutomaticDimension
-    }
-
     override func viewDidLayoutSubviews() {
-        self.tableView.tableHeaderView?.backgroundColor = Theme.backgroundColor
-        self.tableView.tableFooterView?.backgroundColor = Theme.backgroundColor
-        self.tableView.backgroundColor = Theme.backgroundColor
-    }
-    
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-    }
-    
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = self.tableView.dequeueReusableCell(withIdentifier: "mentorFormCell") as! MentorFormCell
-
-        cell.separatorInset = .zero
+        super.viewDidLayoutSubviews()
         
-        return cell
+        descriptionTextView.delegate = self
+        UIView.animate(withDuration: 1.0, animations: {
+            self.view.backgroundColor = Theme.backgroundColor
+            self.contentView.backgroundColor = Theme.backgroundColor
+            self.scrollView.backgroundColor = Theme.backgroundColor
+            
+            self.categoryTextField.attributedPlaceholder = NSAttributedString(string: "Category",
+                                                                              attributes: [NSForegroundColorAttributeName: Theme.tintColor])
+            self.categoryTextField.backgroundColor = Theme.backgroundColor
+            self.categoryTextField.textColor = Theme.primaryColor
+            self.categoryTextField.layer.cornerRadius = 0.0;
+            self.categoryTextField.layer.borderColor = Theme.tintColor.cgColor
+            self.categoryTextField.layer.borderWidth = 1.5
+            
+            self.locationTextField.attributedPlaceholder = NSAttributedString(string: "Location",
+                                                                              attributes: [NSForegroundColorAttributeName: Theme.tintColor])
+            self.locationTextField.backgroundColor = Theme.backgroundColor
+            self.locationTextField.textColor = Theme.primaryColor
+            self.locationTextField.layer.cornerRadius = 0.0;
+            self.locationTextField.layer.borderColor = Theme.tintColor.cgColor
+            self.locationTextField.layer.borderWidth = 1.5
+            
+            self.descriptionTextView.backgroundColor = Theme.backgroundColor
+            self.descriptionTextView.textColor = Theme.tintColor // Manually set the textColor to placeholder
+            self.descriptionTextView.layer.cornerRadius = 0.0;
+            self.descriptionTextView.layer.borderColor = Theme.tintColor.cgColor
+            self.descriptionTextView.layer.borderWidth = 1.5
+            
+            let loginButtonAttributedTitle = NSAttributedString(string: "Submit Ticket",
+                                                                attributes: [NSForegroundColorAttributeName : Theme.primaryColor])
+            self.submitTicketButton.setAttributedTitle(loginButtonAttributedTitle, for: .normal)
+            self.submitTicketButton.layer.cornerRadius = 0.0;
+            self.submitTicketButton.layer.borderColor = Theme.tintColor.cgColor
+            self.submitTicketButton.layer.borderWidth = 1.5
+        })
+
+        
     }
     
-    override func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        let headerCell = self.tableView.dequeueReusableCell(withIdentifier: "headerCell") as! SpartaTableViewHeaderCell
-        headerCell.separatorInset = .zero
-        headerCell.titleLabel.text = "Create A Ticket"
-        return headerCell
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        textField.resignFirstResponder()
+        return true
     }
     
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 1
+    func keyboardWillShow(notification:NSNotification) {
+        scrollViewContentInset = self.scrollView.contentInset // save the current content inset to use on keyboardWillHide
+        // (We need to save the original content inset because our custom navigation bar already set it
+        var userInfo = notification.userInfo!
+        var keyboardFrame: CGRect = (userInfo[UIKeyboardFrameBeginUserInfoKey] as! NSValue).cgRectValue
+        keyboardFrame = self.view.convert(keyboardFrame, from: nil)
+        
+        var contentInset: UIEdgeInsets = self.scrollView.contentInset
+        contentInset.bottom = keyboardFrame.size.height
+        self.scrollView.contentInset = contentInset
     }
     
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return 1
+    func keyboardWillHide(notification:NSNotification){
+        self.scrollView.contentInset = scrollViewContentInset
     }
     
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        self.tableView.deselectRow(at: indexPath, animated: true)
-    }
-    
-    override func requiresLogin() -> Bool {
+    func requiresLogin() -> Bool {
         return true
     }
